Make `setadbackend` more consistent + use ZygoteRules
