+++ src/core/Core.jl
@@ -15,13 +15,13 @@ using StatsFuns: logsumexp, softmax
 @reexport using DynamicPPL
 using Requires
 
+import ZygoteRules
+
 include("container.jl")
 include("ad.jl")
+include("deprecations.jl")
+
 function __init__()
-    @require Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f" begin
-        include("compat/zygote.jl")
-        export ZygoteAD
-    end
     @require ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267" begin
         include("compat/reversediff.jl")
         export ReverseDiffAD, getrdcache, setrdcache, emptyrdcache
@@ -51,6 +51,7 @@ export  @model,
         setadsafe,
         ForwardDiffAD,
         TrackerAD,
+        ZygoteAD,
         value,
         gradient_logp,
         CHUNKSIZE,
